spike-37: add config data to config file
