+++ nest-app/package.json
@@ -29,6 +29,7 @@
     "class-sanitizer": "^1.0.1",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
+    "dotenv": "^16.0.3",
     "http-status-codes": "^2.2.0",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.2.0"
