+++ nest-app/src/app.module.ts
@@ -4,6 +4,7 @@ import { AccountModule } from "./account/account.module";
 import { AppController } from "./app.controller";
 import { AppService } from "./app.service";
 import { BlocksModule } from "./blocks/blocks.module";
+import { NODE_CACHE_LIMIT } from "./config";
 import { DeploysModule } from "./deploys/deploys.module";
 import { PeersModule } from "./peers/peers.module";
 import { ValidatorsModule } from "./validators/validators.module";
@@ -16,7 +17,7 @@ import { ValidatorsModule } from "./validators/validators.module";
     DeploysModule,
     AccountModule,
     ScheduleModule.forRoot(),
-    CacheModule.register({ ttl: Number.MAX_VALUE, max: 2 ** 19 }),
+    CacheModule.register({ ttl: Number.MAX_VALUE, max: NODE_CACHE_LIMIT }),
   ],
   controllers: [AppController],
   providers: [AppService],
