+++ nest-app/src/blocks/blocks.module.ts
@@ -1,9 +1,12 @@
 import { CacheModule, Module } from "@nestjs/common";
+import { NODE_CACHE_LIMIT } from "src/config";
 import { BlocksController } from "./blocks.controller";
 import { BlocksService } from "./blocks.service";
 
 @Module({
-  imports: [CacheModule.register({ ttl: Number.MAX_VALUE, max: 2 ** 19 })],
+  imports: [
+    CacheModule.register({ ttl: Number.MAX_VALUE, max: NODE_CACHE_LIMIT }),
+  ],
   controllers: [BlocksController],
   providers: [BlocksService],
   exports: [BlocksService],
