--- nest-app/src/types/peers.ts
@@ -1,7 +1,0 @@
-export interface Peer {
-  nodeId: string;
-  address: string;
-  isAlive?: boolean | null;
-  uptime?: string | null;
-  lastAddedBlockHash?: string | null;
-}
