+++ nest-app/src/validators/validators.module.ts
@@ -1,10 +1,13 @@
 import { CacheModule, Module } from "@nestjs/common";
 import { BlocksService } from "src/blocks/blocks.service";
+import { NODE_CACHE_LIMIT } from "src/config";
 import { ValidatorsController } from "./validators.controller";
 import { ValidatorsService } from "./validators.service";
 
 @Module({
-  imports: [CacheModule.register({ ttl: Number.MAX_VALUE, max: 2 ** 19 })],
+  imports: [
+    CacheModule.register({ ttl: Number.MAX_VALUE, max: NODE_CACHE_LIMIT }),
+  ],
   controllers: [ValidatorsController],
   providers: [ValidatorsService, BlocksService],
 })
